+++ src/test/kotlin/FixedPointTest.kt
@@ -7,8 +7,8 @@ class FixedPointTest() {
     fun help(matrixS: String, bS: String) {
         val matrix = readMatrixFromFile(matrixS)
         val b = readMatrixFromFile(bS)
-        val answer = solveFixedPoint(matrix, b, Constants.Epsilon.toBigDecimal())
-        assert((answer.matrix - matrix * answer.matrix - b).norm() <= Constants.Scale.toBigDecimal())
+        val answer = solveFixedPoint(matrix, b, Constants.Epsilon)
+        assert((answer.matrix - matrix * answer.matrix - b).norm() <= Constants.Scale)
     }
 
     @Test
